Fixed the stages system so that new stages are generated as the program runs (when one stage is in progress the next stage is generated), rather than generating all the stages at startup. This should conserve RAM.
Now I have to figure out initial stages and unit figures that will make the drones' economy and army run smootly.

Or, my dream is to create an adaptive algorithm that will let the program figure out its own ideal values. That would be a kick-ass science fair project, because apparently they accept computer science projects. I want to be able to go to the likes of MIT or Caltech, but those are competitive admissions and boy would a nice project give me a leg up.....

Anyway, two commits in one day. I am on FIRE!!!!!